break error defs into error.rs; update registry client
